+++ mitra_adapters/src/payments/subscriptions.rs
@@ -1,3 +1,5 @@
+use std::num::NonZeroU64;
+
 use chrono::{DateTime, Duration, Utc};
 
 use mitra_models::{
@@ -11,6 +13,24 @@ use mitra_models::{
     subscriptions::types::DbSubscription,
     users::types::User,
 };
+use mitra_validators::errors::ValidationError;
+
+// TODO: should be based on network fee
+const MONERO_PAYMENT_AMOUNT_MIN: u64 = 1000000000; // 0.001
+
+const SECONDS_IN_MONTH: u64 = 30 * 24 * 3600;
+
+pub fn validate_subscription_price(
+    value: NonZeroU64,
+) -> Result<(), ValidationError> {
+    let price_per_second = value.get();
+    let price_per_month = price_per_second.checked_mul(SECONDS_IN_MONTH)
+        .ok_or(ValidationError("price is too high"))?;
+    if price_per_month < MONERO_PAYMENT_AMOUNT_MIN {
+        return Err(ValidationError("price is too low"));
+    };
+    Ok(())
+}
 
 pub async fn create_or_update_subscription(
     db_client: &mut impl DatabaseClient,
@@ -80,4 +100,23 @@ pub async fn create_or_update_local_subscription(
             }
         },
     ).await
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_validate_subscription_price() {
+        let price = NonZeroU64::new(500).unwrap();
+        let result = validate_subscription_price(price);
+        assert_eq!(result.is_ok(), true);
+    }
+
+    #[test]
+    fn test_validate_subscription_price_too_small() {
+        let price = NonZeroU64::new(350).unwrap();
+        let result = validate_subscription_price(price);
+        assert_eq!(result.is_ok(), false);
+    }
 }