+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Write message to log before forwarding payment.
+- Require subscription price to be greater than minimum payment amount.
 
 ### Fixed